refactor: move project type to config
