+++ src/config/project.ts
@@ -1,0 +1,8 @@
+export type Project = {
+    url: string;
+    slug: string;
+    title: string;
+    tags: string[];
+    description: string;
+    publishedDate: string;
+};
